+++ packages/manager/apps/freefax/src/index.js
@@ -2,6 +2,7 @@ import 'script-loader!jquery'; // eslint-disable-line
 import 'script-loader!lodash'; // eslint-disable-line
 
 import angular from 'angular';
-import '@ovh-ux/manager-freefax';
+import ngOvhApiWrappers from '@ovh-ux/ng-ovh-api-wrappers';
+import ovhManagerFreeFax from '@ovh-ux/manager-freefax';
 
-angular.module('freefaxApp', ['ovhManagerFreeFax']);
+angular.module('freefaxApp', [ngOvhApiWrappers, ovhManagerFreeFax]);