+++ packages/manager/apps/iplb/src/index.js
@@ -6,7 +6,9 @@ import 'script-loader!moment/min/moment-with-locales.min';
 import { Environment } from '@ovh-ux/manager-config';
 
 import angular from 'angular';
-import '@ovh-ux/manager-iplb';
+import ngOvhCloudUniverseComponents from '@ovh-ux/ng-ovh-cloud-universe-components';
+import ovhManagerCore from '@ovh-ux/manager-core';
+import ovhManagerIplb from '@ovh-ux/manager-iplb';
 
 import { momentConfiguration } from './config';
 
@@ -15,7 +17,11 @@ import 'ovh-ui-kit-bs/dist/ovh-ui-kit-bs.css';
 Environment.setRegion(__WEBPACK_REGION__);
 
 angular
-  .module('iplbApp', ['ovhManagerIplb'])
+  .module('iplbApp', [
+    ngOvhCloudUniverseComponents,
+    ovhManagerCore,
+    ovhManagerIplb,
+  ])
   .config(
     /* @ngInject */ (CucConfigProvider, coreConfigProvider) => {
       CucConfigProvider.setRegion(coreConfigProvider.getRegion());