+++ packages/manager/modules/emailpro/src/license/history/emailpro-license-history.controller.js
@@ -1,6 +1,6 @@
 import angular from 'angular';
 import set from 'lodash/set';
-import moment from 'moment';
+import 'moment';
 
 export default /* @ngInject */ (
   $rootScope,