+++ packages/manager/modules/vps/src/dashboard/modal/terminate-option/vps-option-terminate.controller.js
@@ -1,3 +1,5 @@
+import 'moment';
+
 export default class VpsOptionTerminateCtrl {
   /* @ngInject */
   constructor($translate, CucControllerHelper, CucCloudMessage, VpsService) {