+++ packages/manager/modules/vps/src/secondary-dns/vps-secondary-dns.controller.js
@@ -1,4 +1,5 @@
 import map from 'lodash/map';
+import 'moment';
 
 export default class {
   /* @ngInject */