+++ packages/manager/apps/vps/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [0.2.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-vps-app@0.1.4...@ovh-ux/manager-vps-app@0.2.0) (2020-04-15)
+
+
+### Features
+
+* **vps:** lazy load vps root state ([28cc74e](https://github.com/ovh/manager/commit/28cc74ee6c5f8789892978fc297ba47b0a2a20ea))
+
+
+### Performance Improvements
+
+* import dynamically moment locales ([#2755](https://github.com/ovh/manager/issues/2755)) ([5f3320d](https://github.com/ovh/manager/commit/5f3320d92802a1f4a6d65baf60f74917b8e58f4a))
+
+
+
 ## [0.1.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-vps-app@0.1.3...@ovh-ux/manager-vps-app@0.1.4) (2020-03-18)