+++ packages/manager/modules/cloud-universe-components/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.7.3](https://github.com/ovh/manager/compare/@ovh-ux/ng-ovh-cloud-universe-components@1.7.2...@ovh-ux/ng-ovh-cloud-universe-components@1.7.3) (2020-04-15)
+
+
+### Performance Improvements
+
+* import dynamically moment locales ([#2755](https://github.com/ovh/manager/issues/2755)) ([5f3320d](https://github.com/ovh/manager/commit/5f3320d92802a1f4a6d65baf60f74917b8e58f4a))
+
+
+
 ## [1.7.2](https://github.com/ovh/manager/compare/@ovh-ux/ng-ovh-cloud-universe-components@1.7.1...@ovh-ux/ng-ovh-cloud-universe-components@1.7.2) (2020-01-22)