+++ packages/manager/modules/exchange/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.8.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-exchange@1.7.1...@ovh-ux/manager-exchange@1.8.0) (2020-04-15)
+
+
+### Features
+
+* **exchange:** lazy load exchange root state ([63e6bbb](https://github.com/ovh/manager/commit/63e6bbb23b14ea820d0f50500db9b881d12775a6))
+
+
+### Performance Improvements
+
+* import dynamically moment locales ([#2755](https://github.com/ovh/manager/issues/2755)) ([5f3320d](https://github.com/ovh/manager/commit/5f3320d92802a1f4a6d65baf60f74917b8e58f4a))
+
+
+
 ## [1.7.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-exchange@1.7.0...@ovh-ux/manager-exchange@1.7.1) (2020-04-08)